smart-student-assitant/requirements.txt
python-dotenv
chainlit
google-generativeai==0.8.5
